+++ src/constants/errors.ts
@@ -1,0 +1,5 @@
+export enum Errors {
+  MissingApiKey = 'MissingApiKey',
+  InvalidApiKey = 'InvalidApiKey',
+  DeployFailed = 'DeployFailed',
+}