+++ webpack.config.js
@@ -1,0 +1,31 @@
+const path = require('path');
+const webpack = require('webpack');
+
+module.exports = {
+  mode: "development",
+  devtool: "inline-source-map",
+  entry: {
+    main: "./src/index.ts",
+  },
+  output: {
+    library: "FlexSmart",
+    path: path.resolve(__dirname, './dist_bundle'),
+    filename: "bundle.js"
+  },
+  resolve: {
+    extensions: [".ts", ".tsx", ".js"],
+  },
+  plugins: [
+    new webpack.DefinePlugin({
+      'process.env.NODE_DEBUG': JSON.stringify(process.env.NODE_DEBUG),
+    }),
+  ],
+  module: {
+    rules: [
+      { 
+        test: /\.tsx?$/,
+        loader: "ts-loader"
+      }
+    ]
+  }
+};